refactor: move prepare_table_with_two_btree_indexes to data.rs
- Move test data preparation function from table_update.rs to
  integration-tests/src/data.rs alongside other prepare_* helpers
- Add indexlake-index-btree to integration-tests dependencies
- Update tests to import from indexlake_integration_tests::data
- Return ILResult<Table> instead of tuple, tests get catalog via client.catalog.clone()
