+++ integration-tests/src/data.rs
@@ -6,9 +6,11 @@ use arrow::array::{
 use arrow::datatypes::{DataType, Field, Schema};
 use geo::{Geometry, Point};
 use geozero::{CoordDimensions, ToWkb};
+use indexlake::index::IndexKind;
 use indexlake::storage::DataFileFormat;
-use indexlake::table::{Table, TableConfig, TableCreation, TableInsertion};
+use indexlake::table::{IndexCreation, Table, TableConfig, TableCreation, TableInsertion};
 use indexlake::{Client, ILResult};
+use indexlake_index_btree::{BTreeIndexKind, BTreeIndexParams};
 
 use crate::utils::{assert_data_file_count, assert_inline_row_count};
 
@@ -360,4 +362,92 @@ pub async fn prepare_simple_vector_table(
     assert_data_file_count(&table, |count| count > 0).await?;
 
     Ok(table)
+}
+
+pub async fn prepare_table_with_two_btree_indexes(
+    client: &Client,
+    format: DataFileFormat,
+) -> ILResult<Table> {
+    let namespace_name = uuid::Uuid::new_v4().to_string();
+    client.create_namespace(&namespace_name, true).await?;
+
+    let list_inner_field = Arc::new(Field::new("item", DataType::Float32, false));
+    let table_schema = Arc::new(Schema::new(vec![
+        Field::new("id", DataType::Int32, false),
+        Field::new("name", DataType::Utf8, false),
+        Field::new("age", DataType::Int32, false),
+        Field::new_list(
+            "vector",
+            Arc::new(Field::new("item", DataType::Float32, false)),
+            false,
+        ),
+    ]));
+    // Keep rows inline so inline indexes are used
+    let table_config = TableConfig {
+        inline_row_count_limit: 100,
+        parquet_row_group_size: 2,
+        preferred_data_file_format: format,
+    };
+    let table_name = uuid::Uuid::new_v4().to_string();
+    let table_creation = TableCreation {
+        namespace_name: namespace_name.clone(),
+        table_name: table_name.clone(),
+        schema: table_schema.clone(),
+        config: table_config,
+        ..Default::default()
+    };
+    client.create_table(table_creation).await?;
+    let table = client.load_table(&namespace_name, &table_name).await?;
+
+    let mut vector_builder = ListBuilder::new(Float32Builder::new()).with_field(list_inner_field);
+    for vector in [
+        [10.0, 10.0, 10.0],
+        [20.0, 20.0, 20.0],
+        [30.0, 30.0, 30.0],
+        [40.0, 40.0, 40.0],
+    ] {
+        vector_builder.values().append_slice(&vector);
+        vector_builder.append(true);
+    }
+
+    let record_batch = RecordBatch::try_new(
+        table_schema.clone(),
+        vec![
+            Arc::new(Int32Array::from(vec![1, 2, 3, 4])),
+            Arc::new(StringArray::from(vec!["Alice", "Bob", "Charlie", "David"])),
+            Arc::new(Int32Array::from(vec![20, 21, 22, 23])),
+            Arc::new(vector_builder.finish()),
+        ],
+    )?;
+    table
+        .insert(TableInsertion::new(vec![record_batch]))
+        .await?;
+
+    // Create btree index on name
+    let name_index_creation = IndexCreation {
+        name: "name_btree".to_string(),
+        kind: BTreeIndexKind.kind().to_string(),
+        key_columns: vec!["name".to_string()],
+        params: Arc::new(BTreeIndexParams {}),
+        concurrency: 1,
+        if_not_exists: false,
+    };
+    table.create_index(name_index_creation).await?;
+
+    // Reload table after create_index consumes self
+    let table = client.load_table(&namespace_name, &table_name).await?;
+
+    // Create btree index on age
+    let age_index_creation = IndexCreation {
+        name: "age_btree".to_string(),
+        kind: BTreeIndexKind.kind().to_string(),
+        key_columns: vec!["age".to_string()],
+        params: Arc::new(BTreeIndexParams {}),
+        concurrency: 1,
+        if_not_exists: false,
+    };
+    table.create_index(age_index_creation).await?;
+
+    // Reload table after create_index consumes self
+    client.load_table(&namespace_name, &table_name).await
 }
