+++ integration-tests/tests/table_update.rs
@@ -1,18 +1,15 @@
-use arrow::array::{Float32Builder, Int32Array, ListBuilder, RecordBatch, StringArray};
-use arrow::datatypes::{DataType, Field, Schema};
+use arrow::array::{Float32Builder, ListBuilder};
+use arrow::datatypes::{DataType, Field};
 use indexlake::Client;
 use indexlake::catalog::{
     Catalog, CatalogDataType, CatalogSchema, Column, INTERNAL_ROW_ID_FIELD_NAME, Scalar,
 };
 use indexlake::expr::{col, lit};
-use indexlake::index::IndexKind;
 use indexlake::storage::{DataFileFormat, Storage};
-use indexlake::table::{
-    IndexCreation, TableConfig, TableCreation, TableInsertion, TableScan, TableUpdate,
-};
-use indexlake_index_btree::{BTreeIndexKind, BTreeIndexParams};
+use indexlake::table::{TableScan, TableUpdate};
+use indexlake_index_btree::BTreeIndexKind;
 use indexlake_integration_tests::data::{
-    prepare_simple_testing_table, prepare_simple_vector_table,
+    prepare_simple_testing_table, prepare_simple_vector_table, prepare_table_with_two_btree_indexes,
 };
 use indexlake_integration_tests::utils::{
     full_table_scan, read_first_row_id_bytes_from_table, table_scan,
@@ -166,95 +163,6 @@ async fn update_table_by_catalog_unsupported_condition(
     Ok(())
 }
 
-async fn prepare_table_with_two_btree_indexes(
-    client: &Client,
-    format: DataFileFormat,
-) -> Result<(indexlake::table::Table, Arc<dyn Catalog>), Box<dyn std::error::Error>> {
-    let namespace_name = uuid::Uuid::new_v4().to_string();
-    client.create_namespace(&namespace_name, true).await?;
-
-    let table_schema = Arc::new(Schema::new(vec![
-        Field::new("id", DataType::Int32, false),
-        Field::new("name", DataType::Utf8, false),
-        Field::new("age", DataType::Int32, false),
-        Field::new_list(
-            "vector",
-            Arc::new(Field::new("item", DataType::Float32, false)),
-            false,
-        ),
-    ]));
-    // Keep rows inline so inline indexes are used
-    let table_config = TableConfig {
-        inline_row_count_limit: 100,
-        parquet_row_group_size: 2,
-        preferred_data_file_format: format,
-    };
-    let table_name = uuid::Uuid::new_v4().to_string();
-    let table_creation = TableCreation {
-        namespace_name: namespace_name.clone(),
-        table_name: table_name.clone(),
-        schema: table_schema.clone(),
-        config: table_config,
-        ..Default::default()
-    };
-    client.create_table(table_creation).await?;
-    let table = client.load_table(&namespace_name, &table_name).await?;
-
-    let list_inner_field = Arc::new(Field::new("item", DataType::Float32, false));
-    let mut vector_builder = ListBuilder::new(Float32Builder::new()).with_field(list_inner_field);
-    for vector in [
-        [10.0, 10.0, 10.0],
-        [20.0, 20.0, 20.0],
-        [30.0, 30.0, 30.0],
-        [40.0, 40.0, 40.0],
-    ] {
-        vector_builder.values().append_slice(&vector);
-        vector_builder.append(true);
-    }
-
-    let record_batch = RecordBatch::try_new(
-        table_schema.clone(),
-        vec![
-            Arc::new(Int32Array::from(vec![1, 2, 3, 4])),
-            Arc::new(StringArray::from(vec!["Alice", "Bob", "Charlie", "David"])),
-            Arc::new(Int32Array::from(vec![20, 21, 22, 23])),
-            Arc::new(vector_builder.finish()),
-        ],
-    )?;
-    table
-        .insert(TableInsertion::new(vec![record_batch]))
-        .await?;
-
-    // Create btree index on name
-    let name_index_creation = IndexCreation {
-        name: "name_btree".to_string(),
-        kind: BTreeIndexKind.kind().to_string(),
-        key_columns: vec!["name".to_string()],
-        params: Arc::new(BTreeIndexParams {}),
-        concurrency: 1,
-        if_not_exists: false,
-    };
-    table.create_index(name_index_creation).await?;
-
-    // Reload table after create_index consumes self
-    let table = client.load_table(&namespace_name, &table_name).await?;
-
-    // Create btree index on age
-    let age_index_creation = IndexCreation {
-        name: "age_btree".to_string(),
-        kind: BTreeIndexKind.kind().to_string(),
-        key_columns: vec!["age".to_string()],
-        params: Arc::new(BTreeIndexParams {}),
-        concurrency: 1,
-        if_not_exists: false,
-    };
-    table.create_index(age_index_creation).await?;
-
-    // Reload table after create_index consumes self
-    let table = client.load_table(&namespace_name, &table_name).await?;
-    Ok((table, client.catalog.clone()))
-}
-
 async fn count_inline_index_records(
     catalog: Arc<dyn Catalog>,
     table_id: &Uuid,
@@ -323,7 +231,8 @@ async fn update_non_index_column_does_not_affect_indexes(
     let mut client = Client::new(catalog, storage);
     client.register_index_kind(Arc::new(BTreeIndexKind));
 
-    let (table, catalog) = prepare_table_with_two_btree_indexes(&client, format).await?;
+    let table = prepare_table_with_two_btree_indexes(&client, format).await?;
+    let catalog = client.catalog.clone();
     assert_eq!(
         count_inline_index_records(catalog.clone(), &table.table_id).await?,
         2
@@ -400,7 +309,8 @@ async fn update_one_index_column_only_affects_that_index(
     let mut client = Client::new(catalog, storage);
     client.register_index_kind(Arc::new(BTreeIndexKind));
 
-    let (table, catalog) = prepare_table_with_two_btree_indexes(&client, format).await?;
+    let table = prepare_table_with_two_btree_indexes(&client, format).await?;
+    let catalog = client.catalog.clone();
     assert_eq!(
         count_inline_index_records(catalog.clone(), &table.table_id).await?,
         2
